Move `EntryId` to `re_types_core`
