+++ rerun_py/src/catalog/entry.rs
@@ -2,8 +2,8 @@ use std::str::FromStr as _;
 
 use pyo3::{exceptions::PyTypeError, pyclass, pymethods, Py, PyErr, PyResult, Python};
 
+use re_log_types::EntryId;
 use re_protos::catalog::v1alpha1::{ext::EntryDetails, EntryKind};
-use re_protos::common::v1alpha1::ext::EntryId;
 
 use crate::catalog::PyCatalogClient;
